replace `hyper` v0.10  with `ureq`
`hyper` v0.10 is extremely outdated, the current release is v1.6. this
old version pulls in vulnerable dependencies (incl. `hyper` v0.10
itself).

rather than upgrading to `hyper` v1.6 i opted to replace it with
`ureq` since this crate here is not `async` and `hyper` v1 is completely
`async`. due to the very limited amount of features needed from `hyper`
(just executing HTTP POST requests) it can instead be replaced with the
much simpler `ureq` crate.

i initially opted for `reqwest`, however `reqwest` is primarily `async`
(and uses `hyper` underneath). while it offers a `blocking` layer this
causes major issues when integrating it further up the stack in an
`async` context as also stated by the [`reqwest::blocking` docs].

this is a breaking change for consumers since the URLs can now just be
passed as strings rather than having to call `.parse()?` on the string.

note that this is similar to, but not the same as, #8 and #11.

fixes #9

[`reqwest::blocking` docs]: https://docs.rs/reqwest/latest/reqwest/blocking/